+++ sdk/resourcemanager/operationalinsights/armoperationalinsights/ze_generated_example_datasources_client_test.go
@@ -100,7 +100,7 @@ func ExampleDataSourcesClient_Get() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/operationalinsights/resource-manager/Microsoft.OperationalInsights/stable/2020-08-01/examples/DataSourcesListByWorkspace.json
-func ExampleDataSourcesClient_ListByWorkspace() {
+func ExampleDataSourcesClient_NewListByWorkspacePager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -112,7 +112,7 @@ func ExampleDataSourcesClient_ListByWorkspace() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByWorkspace("<resource-group-name>",
+	pager := client.NewListByWorkspacePager("<resource-group-name>",
 		"<workspace-name>",
 		"<filter>",
 		&armoperationalinsights.DataSourcesClientListByWorkspaceOptions{Skiptoken: nil})