+++ sdk/resourcemanager/operationalinsights/armoperationalinsights/ze_generated_example_storageinsightconfigs_client_test.go
@@ -107,7 +107,7 @@ func ExampleStorageInsightConfigsClient_Delete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/operationalinsights/resource-manager/Microsoft.OperationalInsights/stable/2020-08-01/examples/StorageInsightsListByWorkspace.json
-func ExampleStorageInsightConfigsClient_ListByWorkspace() {
+func ExampleStorageInsightConfigsClient_NewListByWorkspacePager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -119,7 +119,7 @@ func ExampleStorageInsightConfigsClient_ListByWorkspace() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByWorkspace("<resource-group-name>",
+	pager := client.NewListByWorkspacePager("<resource-group-name>",
 		"<workspace-name>",
 		nil)
 	for pager.More() {